my final year project
This is a blockchain-based alcohol provenance system that enhances authenticity in the alcohol industry. This is a QR code based distributed application powered by the Ethereum network. The manufacturers will input bottle information from the web app which will then be stored in the Ethereum blockchain. The consumers can use the QRcode on the bottle to retrieve the information of the bottle from the blockchain.
Technologies used:
Laravel, MySQl, Web3 js, Solidity, Metamask